Batch completed + use of javaMailSender with creation of an account on Ethereal

+++ src/main/java/com/emilie/SpringBatch/Service/LoanStatusService.java
@@ -21,7 +21,7 @@ public class LoanStatusService {
 
 
     public void updateLoanStatus(String accessToken){
-        Integer nbOfUpdateLoan = feignProxy.updateLoanStatus(accessToken);
+        Long nbOfUpdateLoan = feignProxy.updateLoanStatus(accessToken);
     }
 
     public List<Loan> getDelay(String accessToken){
